BSATE-33: add pos_hide_cost_price_and_margin V15.0
This module is to hide cost, margin, and replenishment info from product
detail in pos. It utilize a setting from pos_margin module but doesn't
need to depend on the later module thanks to the mechanism that if there
is no field iface_display_margin, the value of t-if is False.